# api/openapi/xiuxian/v1/world.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "xiuxian/v1/world.proto",
    "version": "version not set"
  },
...
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/xiuxian.v1.WorldService/GetState": {
      "post": {
        "operationId": "WorldService_GetState",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1RoleState"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1GetStateRequest"
            }
          }
        ],
        "tags": [
          "WorldService"
        ]
      }
    },
    "/xiuxian.v1.WorldService/GetWorldBounds": {
      "post": {
        "operationId": "WorldService_GetWorldBounds",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1WorldBounds"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1GetWorldBoundsRequest"
            }
          }
        ],
        "tags": [
          "WorldService"
        ]
      }
    },
...
  },
  "definitions": {
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
...
    "v1GetStateRequest": {
      "type": "object"
    },
    "v1GetWorldBoundsRequest": {
      "type": "object"
    },
...
    "v1Position": {
      "type": "object",
      "properties": {
        "xMilliunits": {
          "type": "string",
          "format": "int64"
        },
        "yMilliunits": {
          "type": "string",
          "format": "int64"
        }
      }
    },
...
    "v1RoleState": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "lifeNumber": {
          "type": "string",
          "format": "int64"
        },
        "status": {
          "type": "string"
        },
        "cultivationMillis": {
          "type": "string",
          "format": "int64"
        },
        "realmLevel": {
          "type": "integer",
          "format": "int32"
        },
        "realmName": {
          "type": "string"
        },
        "ageMillis": {
          "type": "string",
          "format": "int64"
        },
        "lifespanMillis": {
          "type": "string",
          "format": "int64"
        },
        "speed": {
          "type": "string",
          "format": "int64"
        },
        "senseRadius": {
          "type": "string",
          "format": "int64"
        },
        "position": {
          "$ref": "#/definitions/v1Position"
        },
        "movementState": {
          "type": "string"
        },
        "stateVersion": {
          "type": "string",
          "format": "int64"
        }
      }
    },
...
    "v1WorldBounds": {
      "type": "object",
      "properties": {
        "minXMilliunits": {
          "type": "string",
          "format": "int64"
        },
        "maxXMilliunits": {
          "type": "string",
          "format": "int64"
        },
        "minYMilliunits": {
          "type": "string",
          "format": "int64"
        },
        "maxYMilliunits": {
          "type": "string",
          "format": "int64"
        }
      }
    },
...
  }
}